First functional version of the TargetResolver class (it works!)
After a few hours tinkering with the SIMBAD catalog we finally have
a functional version of the code: TargetResolver.resolve(String
targetName) receives the name of the object, queries the database
and returns a TargetInformation instance with encapsulates all the
information.

This is just first base. Next steps should include properly and
thoroughly documenting the code, refactoring it where needed and, most
important, write extensive unit test cases that query the database a
few thousand times, in order to make sure that the Observation Tool
does not crash because of our fault.