Refactoring complete / Resolver now throws exceptions.
Both classes, TargetInformation and TargetResolver, have been
completely refactored and commented -- although not following Javadoc
conventions; which I won't do unless told so by A. Garcia. The code
should now be robust enough to be straight-forwardly included in the
Observation Tool or, at least, please my fellow developers before they
politely ask me to change this and that.

The TargetResolver now throws two exceptions: SIMBADQueryException if
the connection to SIMBAD fails and TargetNotFoundException when the
object cannot be resolved by (that is, found in) SIMBAD. No need for
more null checking this: if no exception is thrown, it means that the
connection was successfully established and the object resolved.