Added unit tests using JUnit.
Tests for for both classes, TargetInformation and TargetResolver, have
been written using the JUnit framework. For the most important method
(and the only one A. Garcia the PANIC Observation Tool will need to
invoke), TargetResolver.submit(String), a total of 1030 different (and
actual) objects are attempted to be resolved, plus a few randomly
generated hexadecimal strings to test for objects that cannot be found
in the database. Each of these targets is resolved twice: once with
the default TargetResolver instance (ICRS, epoch and equinox = 2000),
and the other one with random values (SIMBAD accepts six different
celestial reference systems: ICRS, FK5, FK4, GAL, SGAL and ECL) and a
random year between zero and the present one.

Speaking of celestial reference systems, by the way: they are no
longer stored in a String data type. Instead, the six allowed values
are encapsulated now in an Enum class, which is passed to the
TargetResolver, so that no invalid reference system can ever be used.