Timeout added to the connection to SIMBAD.
The query to SIMBAD is now run in a different thread, using the
Callable interface. The connection will be aborted and the
SIMBADQueryException thrown in case the timeout value (currently five
seconds, as suggested by José Miguel Ibáñez) is exceeded. The SIMBAD
database is reliable enough as to be expected to be online at all
times, but at the very least this will prevent our code from hanging
up in case network access is lost -- something Antonio García
experienced a few days ago and the reason why this modification has
been made.

Also, and as we discussed at our last PANIC video-conference, the
epoch value used by the Target Resolver has been changed to the
current year, while the equinox is still set to J2000. In this manner,
our coordinates have always their proper motions corrected, something
that was not trivial (if even possible) to achieve with GEIRS.