Add the two fundamental methods of the Staralt class.
I've been asked to also implement a "staralt" (generator of graphs
which plot the altitude of an astronomical object against time), which
will be part of the PANIC observation tool. My workmates and immediate
superiors agree that not reinventing the wheel and standing on giant's
shoulders is the way to go: in this case, it means reusing Marco
Azzaro's superb Staralt (http://catserver.ing.iac.es/staralt/)

Although not free software, Marco kindly shared the source code with
us and allowed it to be used in the observation tool. It is, alas,
written in Fortran, which severely conflicts with our intention to
develop a pure Java application. The accepted solution is to make
use of the online version of the program to download the plot.

The first method of the class, build_query_url, returns the URL of the
plot generated by Staralt for an object and date. Although PANIC will
be commissioned at Calar Alto, the coordinates of the observatory and
its altitude must be given to the method so that the class can be used
anywhere. The second method, download, receives the URL of an online
Staralt plot and saves it to the default temporary-file directory,
returning its absolute path.