Add class constructors.
The coordinates and height of the observatory will not have to be
given by the user every time an object is plotted, but instead only
defined when the Staralt object is created. Without arguments, the
class defaults to the coordinates of the Calar Alto Observatory,
where PANIC is expected to be commissioned by the end of this year.