Return a File object instead of a String with the path.
Returning a File object which encapsulates the GIF file to which the
plot is saved is preferable to simply returning its absolute path as a
String. The File will probably be necessary for almost any operation
on the file (such as deleting it when we are done) and, after all, we
were already using a File instance internally.