Do not get sexagesimal coordinates from SIMBAD
The PANIC Observation Tool needs the sexagesimal coordinates to always
be formatted in the same way: a string like "14 45 29.32" in the case
of right ascensions and like "+39 24 58.15" for declinations. However,
we have noticed that, sometimes (for example, for object M52), SIMBAD
returns a string without seconds and where minutes are expressed with
decimals — such as "+61 35.6". Therefore, we need to not query SIMBAD
for the coordinates in sexagesimal and, in order to have control over
how they are formatted, do the conversion ourselves using the
SIMBADQuerier.DD_to_DMS and DD_to_HMS methods.