Take into account that both the RA and DEC can be null
The right ascension and declination of the object cannot be directly
passed to SIMBADQuerier.DD_to_HMS and DD_to_DMS, respectively, since
in some rare cases the coordinates may be unknown to SIMBAD (try, for
example, CoRoT-21b) and therefore SIMBADQuerier would return null. We
need, thus, to check that the coordinates are not null before doing
the conversion to sexagesimal, or the java.lang.NullPointerException
exception may be thrown.